put xrf model in a separate file
