+++ skxray/fitting/models.py
@@ -5,12 +5,6 @@
 # @author: Li Li (lili@bnl.gov)                                        #
 # created on 09/10/2014                                                #
 #                                                                      #
-# Original code:                                                       #
-# @author: Mirna Lerotic, 2nd Look Consulting                          #
-#         http://www.2ndlookconsulting.com/                            #
-# Copyright (c) 2013, Stefan Vogt, Argonne National Laboratory         #
-# All rights reserved.                                                 #
-#                                                                      #
 # Redistribution and use in source and binary forms, with or without   #
 # modification, are permitted provided that the following conditions   #
 # are met:                                                             #
@@ -193,278 +187,3 @@ class Lorentzian2Model(Model):
 
     def __init__(self, *args, **kwargs):
         super(Lorentzian2Model, self).__init__(lorentzian2, *args, **kwargs)
-
-
-
-def gauss_peak_xrf(x, area, center, sigma,
-                   ratio, fwhm_offset, fwhm_fanoprime,
-                   epsilon=2.96):
-    """
-    This is a function to construct xrf element peak, which is based on gauss_peak,
-    but more specific requirements need to be considered.
-
-    Parameters
-    ----------
-    x : array
-        independent variable
-    area : float
-        area of gaussian function
-    center : float
-        center position
-    sigma : float
-        standard deviation
-    ratio : float
-        value used to adjust peak height
-    fwhm_offset : float
-        global fitting parameter for peak width
-    fwhm_fanoprime : float
-        global fitting parameter for peak width
-
-    Returns
-    -------
-    array:
-        gaussian peak profile
-    """
-    def get_sigma(center):
-        temp_val = 2 * np.sqrt(2 * np.log(2))
-        return np.sqrt((fwhm_offset/temp_val)**2 + center*epsilon*fwhm_fanoprime)
-
-    return gauss_peak(x, area, center, sigma*get_sigma(center)) * ratio
-
-
-class GaussModel_xrf(Model):
-
-    __doc__ = _gen_class_docs(gauss_peak_xrf)
-
-    def __init__(self, *args, **kwargs):
-        super(GaussModel_xrf, self).__init__(gauss_peak_xrf, *args, **kwargs)
-        self.set_param_hint('epsilon', value=2.96, vary=False)
-
-
-def _set_value(para_name, input_dict, model_name):
-    """
-    Set parameter information to a given model
-
-    Parameter
-    ---------
-    para_name : str
-        parameter used for fitting
-    input_dict : dict
-        all the initial values and constraints for given parameters
-    model_name : object
-        model object used in lmfit
-    """
-
-    if input_dict['bound_type'] == 'none':
-        model_name.set_param_hint(name=para_name, value=input_dict['value'], vary=True)
-    elif input_dict['bound_type'] == 'fixed':
-        model_name.set_param_hint(name=para_name, value=input_dict['value'], vary=False)
-    elif input_dict['bound_type'] == 'lohi':
-        model_name.set_param_hint(name=para_name, value=input_dict['value'], vary=True,
-                                  min=input_dict['min'], max=input_dict['max'])
-    elif input_dict['bound_type'] == 'lo':
-        model_name.set_param_hint(name=para_name, value=input_dict['value'], vary=True,
-                                  min=input_dict['min'])
-    elif input_dict['bound_type'] == 'hi':
-        model_name.set_param_hint(name=para_name, value=input_dict['value'], vary=True,
-                                  min=input_dict['max'])
-    else:
-        raise ValueError("could not set values for {0}".format(para_name))
-    return
-
-
-class ModelSpectrum(object):
-
-    def __init__(self, config_file='xrf_paramter.json'):
-        """
-        Parameter
-        ---------
-        config_file : str
-            file save all the fitting parameters
-        """
-        file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)),
-                                 config_file)
-
-        if not os.path.exists(file_path):
-            logger.critical('No configuration file can be found.')
-        else:
-            with open(file_path, 'r') as json_data:
-                self.parameter = json.load(json_data)
-                logger.info('Read data from configuration file {0}.'.format(file_path))
-
-        self.element_list = self.parameter['element_list'].split()
-        self.incident_energy = self.parameter['coherent_sct_energy']['value']
-
-        self.parameter_default = get_para()
-        return
-
-    def setComptonModel(self):
-        """
-        setup parameters related to Compton model
-        """
-        compton = ComptonModel()
-
-        compton_list = ['coherent_sct_energy', 'compton_amplitude',
-                        'compton_angle', 'fwhm_offset', 'fwhm_fanoprime',
-                        'compton_gamma', 'compton_f_tail',
-                        'compton_f_step', 'compton_fwhm_corr',
-                        'compton_hi_gamma', 'compton_hi_f_tail']
-
-        logger.debug('Set up paramters for compton model')
-        for name in compton_list:
-            if name in self.parameter.keys():
-                _set_value(name, self.parameter[name], compton)
-            else:
-                _set_value(name, self.parameter_default[name], compton)
-        logger.debug('Finish setting up paramters for compton model')
-        return compton
-
-    def setElasticModel(self):
-        """
-        setup parameters related to Elastic model
-        """
-        elastic = ElasticModel(prefix='e_')
-
-        item = 'coherent_sct_amplitude'
-        if item in self.parameter.keys():
-            _set_value(item, self.parameter[item], elastic)
-        else:
-            _set_value(item, self.parameter_default[item], elastic)
-
-        logger.debug('Set up paramters for elastic model')
-
-        # set constraints for the following global parameters
-        elastic.set_param_hint(name='fwhm_offset', value=0.1, vary=True, expr='fwhm_offset')
-        elastic.set_param_hint(name='fwhm_fanoprime', value=0.0, vary=True, expr='fwhm_fanoprime')
-        elastic.set_param_hint(name='coherent_sct_energy', value=self.incident_energy,
-                               expr='coherent_sct_energy')
-        logger.debug('Finish setting up paramters for elastic model')
-
-        return elastic
-
-    def model_spectrum(self):
-        """
-        Add all element peaks to the model.
-        """
-        incident_energy = self.incident_energy
-        element_list = self.element_list
-        parameter = self.parameter
-
-        mod = self.setComptonModel() + self.setElasticModel()
-
-        element_adjust = []
-        if parameter.has_key('element'):
-            element_adjust = parameter['element'].keys()
-            logger.info('Those elements need to be adjusted {0}'.format(element_adjust))
-        else:
-            logger.info('No adjustment needs to be considered '
-                        'on the position and width of element peak.')
-
-        for ename in element_list:
-            if ename in k_line:
-                e = Element(ename)
-                if e.cs(incident_energy)['ka1'] == 0:
-                    logger.info('{0} Ka emission line is not activated '
-                                'at this energy {1}'.format(ename, incident_energy))
-                    continue
-
-                logger.debug('Started building element peak for {0}'.format(ename))
-
-                for num, item in enumerate(e.emission_line.all[:4]):
-                    line_name = item[0]
-                    val = item[1]
-
-                    if e.cs(incident_energy)[line_name] == 0:
-                        continue
-
-                    gauss_mod = GaussModel_xrf(prefix=str(ename)+'_'+str(line_name)+'_')
-                    gauss_mod.set_param_hint('fwhm_offset', value=0.1, vary=True, expr='fwhm_offset')
-                    gauss_mod.set_param_hint('fwhm_fanoprime', value=0.1, vary=True, expr='fwhm_fanoprime')
-
-                    if line_name == 'ka1':
-                        gauss_mod.set_param_hint('area', value=100, vary=True, min=0)
-                    else:
-                        gauss_mod.set_param_hint('area', value=100, vary=True, min=0,
-                                                 expr=str(ename)+'_ka1_'+'area')
-                    gauss_mod.set_param_hint('center', value=val, vary=False)
-                    gauss_mod.set_param_hint('sigma', value=1, vary=False)
-                    ratio_v = e.cs(incident_energy)[line_name]/e.cs(incident_energy)['ka1']
-                    gauss_mod.set_param_hint('ratio',
-                                             value=ratio_v, vary=False)
-
-                    if ename in element_adjust:
-                        if parameter['element'][ename].has_key(line_name.lower()+'_position'):
-                            pos_val = parameter['element'][ename][line_name.lower()+'_position']
-                            if pos_val != 0:
-                                gauss_mod.set_param_hint('center', value=val, vary=True,
-                                                         min=val*(1-pos_val), max=val*(1+pos_val))
-                                logger.warning('change element {0} {1} postion '
-                                               'within range {2}'.format(ename, line_name, [-pos_val, pos_val]))
-
-                        if parameter['element'][ename].has_key(line_name.lower()+'_width'):
-                            width_val = parameter['element'][ename][line_name.lower()+'_width']
-                            if width_val != 0:
-                                gauss_mod.set_param_hint('sigma', value=1, vary=True,
-                                                         min=1-width_val, max=1+width_val)
-                                logger.warning('change element {0} {1} peak width '
-                                               'within range {2}'.format(ename, line_name, [-width_val, width_val]))
-
-                    mod = mod + gauss_mod
-                logger.debug('Finished building element peak for {0}'.format(ename))
-
-            elif ename in l_line:
-                ename = ename[:-2]
-                e = Element(ename)
-                if e.cs(incident_energy)['la1'] == 0:
-                    logger.info('{0} La1 emission line is not activated '
-                                'at this energy {1}'.format(ename, incident_energy))
-                    continue
-
-                # L lines
-                #gauss_mod = GaussModel_Llines(prefix=str(ename)+'_l_line_')
-
-                #gauss_mod.set_param_hint('area', value=100, vary=True, min=0)
-                #gauss_mod.set_param_hint('fwhm_offset', value=0.1, vary=True, expr='fwhm_offset')
-                #gauss_mod.set_param_hint('fwhm_fanoprime', value=0.1, vary=True, expr='fwhm_fanoprime')
-
-                for num, item in enumerate(e.emission_line.all[4:-4]):
-
-                    line_name = item[0]
-                    val = item[1]
-
-                    if e.cs(incident_energy)[line_name] == 0:
-                        continue
-
-                    gauss_mod = GaussModel_xrf(prefix=str(ename)+'_'+str(line_name)+'_')
-
-                    gauss_mod.set_param_hint('fwhm_offset', value=0.1, vary=True, expr='fwhm_offset')
-                    gauss_mod.set_param_hint('fwhm_fanoprime', value=0.1, vary=True, expr='fwhm_fanoprime')
-
-                    if line_name == 'la1':
-                        gauss_mod.set_param_hint('area', value=100, vary=True)
-                                             #expr=gauss_mod.prefix+'ratio_val * '+str(ename)+'_la1_'+'area')
-                    else:
-                        gauss_mod.set_param_hint('area', value=100, vary=True,
-                                                 expr=str(ename)+'_la1_'+'area')
-
-                    gauss_mod.set_param_hint('center', value=val, vary=False)
-                    gauss_mod.set_param_hint('sigma', value=1, vary=False)
-                    gauss_mod.set_param_hint('ratio',
-                                             value=e.cs(incident_energy)[line_name]/e.cs(incident_energy)['la1'],
-                                             vary=False)
-
-                    mod = mod + gauss_mod
-
-        self.mod = mod
-        return
-
-    def model_fit(self, x, y, w=None, method='leastsq', **kws):
-        """
-        Parameter
-        ---------
-
-        """
-
-        self.model_spectrum()
-        result = self.mod.fit(y, x=x, weights=w)
-        return result
